# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Mahadev Decoration business operations API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Error {
  error: string;
}

export interface BookingRequest {
  id: number;
  customerName: string;
  phone: string;
  eventType: string;
  eventDate: string;
  venue: string;
  /** @nullable */
  message?: string | null;
  status: string;
  createdAt: string;
  updatedAt: string;
}

export interface BookingRequestInput {
  /** @minLength 2 */
  customerName: string;
  /** @minLength 10 */
  phone: string;
  /** @minLength 2 */
  eventType: string;
  eventDate: string;
  /** @minLength 2 */
  venue: string;
  /** @nullable */
  message?: string | null;
}

export interface BookingStatusUpdate {
  /** @minLength 2 */
  status: string;
  /** @nullable */
  reason?: string | null;
}

export interface PaymentRecord {
  id: number;
  /** @nullable */
  bookingRequestId?: number | null;
  customerName: string;
  eventType: string;
  amount: number;
  status: string;
  paymentDate: string;
  /** @nullable */
  notes?: string | null;
}

export interface PaymentInput {
  /** @nullable */
  bookingRequestId?: number | null;
  /** @minLength 2 */
  customerName: string;
  /** @minLength 2 */
  eventType: string;
  /** @minimum 1 */
  amount: number;
  /** @minLength 2 */
  status: string;
  /** @nullable */
  notes?: string | null;
}

export interface AdminSummary {
  todayEvents: number;
  upcomingEvents: number;
  pendingInquiries: number;
  confirmedBookings: number;
  totalRevenue: number;
  advanceReceived: number;
  pendingPayments: number;
  monthlyGrowth: number;
}

/**
 * Invalid request
 */
export type BadRequestResponse = Error;

/**
 * Resource not found
 */
export type NotFoundResponse = Error;

export type GetAdminBookingsParams = {
q?: string;
};