# lib/api-zod/src/generated/types/adminSummary.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Mahadev Decoration business operations API
 * OpenAPI spec version: 0.1.0
 */

export interface AdminSummary {
  todayEvents: number;
  upcomingEvents: number;
  pendingInquiries: number;
  confirmedBookings: number;
  totalRevenue: number;
  advanceReceived: number;
  pendingPayments: number;
  monthlyGrowth: number;
}