# lib/api-zod/src/generated/types/bookingRequest.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Mahadev Decoration business operations API
 * OpenAPI spec version: 0.1.0
 */

export interface BookingRequest {
  id: number;
  customerName: string;
  phone: string;
  eventType: string;
  eventDate: Date;
  venue: string;
  /** @nullable */
  message?: string | null;
  status: string;
  createdAt: Date;
  updatedAt: Date;
}