# lib/api-zod/src/generated/types/bookingRequestInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Mahadev Decoration business operations API
 * OpenAPI spec version: 0.1.0
 */

export interface BookingRequestInput {
  /** @minLength 2 */
  customerName: string;
  /** @minLength 10 */
  phone: string;
  /** @minLength 2 */
  eventType: string;
  eventDate: Date;
  /** @minLength 2 */
  venue: string;
  /** @nullable */
  message?: string | null;
}