# lib/api-zod/src/generated/types/bookingStatusUpdate.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Mahadev Decoration business operations API
 * OpenAPI spec version: 0.1.0
 */

export interface BookingStatusUpdate {
  /** @minLength 2 */
  status: string;
  /** @nullable */
  reason?: string | null;
}