# lib/api-zod/src/generated/types/paymentInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Mahadev Decoration business operations API
 * OpenAPI spec version: 0.1.0
 */

export interface PaymentInput {
  /** @nullable */
  bookingRequestId?: number | null;
  /** @minLength 2 */
  customerName: string;
  /** @minLength 2 */
  eventType: string;
  /** @minimum 1 */
  amount: number;
  /** @minLength 2 */
  status: string;
  /** @nullable */
  notes?: string | null;
}