# lib/api-zod/src/generated/types/paymentRecord.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Mahadev Decoration business operations API
 * OpenAPI spec version: 0.1.0
 */

export interface PaymentRecord {
  id: number;
  /** @nullable */
  bookingRequestId?: number | null;
  customerName: string;
  eventType: string;
  amount: number;
  status: string;
  paymentDate: Date;
  /** @nullable */
  notes?: string | null;
}